Separated dev and staging index files
