+++ APILambda/test/stagingIndex.ts
@@ -1,22 +1,18 @@
 import awsServerlessExpress from "@vendia/serverless-express"
-import { promiseResult, success } from "TiFBackendUtils"
 import { addBenchmarking, addRoutes, createApp } from "../app.js"
 import { addCognitoTokenVerification } from "../auth.js"
 import { ServerEnvironment } from "../env.js"
 import { addEventToRequest } from "../index.js"
+import { testEnv } from "./devIndex.js"
 
-const testEnv: ServerEnvironment = {
-  environment: "dev",
-  eventStartWindowInHours: 1,
-  maxArrivals: 4,
-  setProfileCreatedAttribute: () => promiseResult(success())
+const stagingEnv: ServerEnvironment = {
+  ...testEnv,
+  environment: "staging"
 }
 
-export const app = createApp()
+const app = createApp()
 addBenchmarking(app)
-addCognitoTokenVerification(app, testEnv)
-addRoutes(app, testEnv)
-
-// for stage tests
+addCognitoTokenVerification(app, stagingEnv)
+addRoutes(app, stagingEnv)
 addEventToRequest(app)
 export const handler = awsServerlessExpress({ app })
